create sui-rpc-resolver
