+++ crates/sui-indexer-alt-framework/src/utils/mod.rs
@@ -2,6 +2,3 @@
 // SPDX-License-Identifier: Apache-2.0
 
 //! Utility functions for the indexer framework.
-
-pub mod event_json;
-pub mod package_store;
